requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==11.0.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
# sqlite3 is built into Python standard library

# LLM providers
google-generativeai==0.3.2
anthropic==0.7.8
httpx==0.25.2

# Data handling
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
playwright==1.40.0

# Development
python-dotenv==1.0.0
black==23.11.0
flake8==6.1.0
